requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.0
certifi==2017.4.17
cffi==1.11.2
chardet==3.0.4
cloudinary==1.10.0
colorama==0.3.9
crypto==1.4.1
cryptography==2.1.4
cycler==0.10.0
decorator==4.0.11
et-xmlfile==1.0.1
future==0.16.0
idna==2.5
jdcal==1.3
jellyfish==0.5.6
libusb1==1.6.4
lxml==3.8.0
matplotlib==2.0.2
mock==2.0.0
Naked==0.1.31
networkx==1.11
ntlm-auth==1.0.6
numpy==1.11.3+mkl
observable==0.3.2
olefile==0.44
opencv-python==3.2.0+contrib
openpyxl==2.4.8
pandas==0.20.1
passlib==1.7.1
pbr==3.1.1
pdfminer.six==20170720
Pillow==4.1.1
probableparsing==0.0.1
protobuf==3.3.0
py==1.4.33
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.5.1
pymongo==3.5.1
pymssql==2.1.3
pyodbc==4.0.16
pyparsing==2.2.0
PyPDF2==1.26.0
pypyodbc==1.3.4
pytesseract==0.1.7
pytest==3.1.1
python-bsonjs==0.2.0
python-crfsuite==0.9.1
python-dateutil==2.6.0
python-docx==0.8.6
python-Levenshtein==0.12.0
python-pptx==0.6.7
pytz==2017.2
pyusb==1.0.0
PyWavelets==0.5.2
PyYAML==3.12
requests==2.18.1
requests-ntlm==1.1.0
scikit-image==0.13.0
scipy==0.19.0
selenium==3.5.0
sharepoint==0.4.2
shellescape==3.4.1
six==1.10.0
tabula-py==1.0.0
tensorflow==1.1.0
Unidecode==1.0.22
urllib3==1.21.1
usaddress==0.5.10
validators==0.12.0
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
Werkzeug==0.12.2
win32core==221.36
win32gui==221.6
windmill==1.6
XlsxWriter==1.0.2
